Fix installer so that it installs theme files with proper ownership
The theme installer script accidentally copied the ownership and permissions
from the source theme files into the installed destination.  This meant that
the user who downloaded the theme would own the installed files -- which
might be a slight security hole since no ordinary user is supposed to own
files in /usr, the likely destination for the installed files.  This
modification ensures that the files will be owned by root and writable only
by root.