Thunar:  Fix icon view rubberband selection opacity
I forgot to style the "rubberband" selection in Thunar.  Before this commit,
when the user dragged over a set of items in icon view in Thunar, the
rubberband selection (which shows the extent of the selection) would be a
totally opaque gray or white, and would completely obscure the selected items.

Thanks to @commandlinegamer for reporting this issue, and also for providing
some code to fix this.  I also modified the Caja rubberbanding code so that
Thunar and Caja can look as similar as possible.