Clean up & add a boatload of icons to the icon theme(s)
This is a positively huge commit (sorry about that), the culmination of
weeks and even months of efforts on the icon themes.  Now the icon themes
are actually half-presentable, IMHO.  It's still not ready for a new
release yet since there is more work to do on the dark icon theme, some
LibreOffice drawing glitches need to be rectified, and I still don't have
screenshots for XFCE (and I need to update the MATE screenshots).  The
release is coming soon though!  Hang on there, and use the master branch
in the meantime!